Initialize the SyncManager when the first Realm is opened instead of when Realm is imported.
