+++ lib/extensions.js
@@ -471,6 +471,29 @@ module.exports = function(realmConstructor) {
                 addSchemaIfNeeded(schema, realmConstructor.Permissions.User);
                 addSchemaIfNeeded(schema, realmConstructor.Subscription.ResultSets);
             },
+
+            // Creates the user agent description for the JS binding itself. Users must specify the application
+            // user agent using Realm.Sync.setUserAgent(...)
+            _createUserAgentDescription() {
+                // Detect if in ReactNative (running on a phone) or in a Node.js environment
+                // Credit: https://stackoverflow.com/questions/39468022/how-do-i-know-if-my-code-is-running-as-react-native
+                try {
+                    var userAgent = "RealmJS/";
+                    var context = getContext();
+                    userAgent = userAgent + require('../package.json').version + " (" + context + ", ";
+                    if (typeof navigator !== 'undefined' && navigator.product === 'ReactNative') {
+                        // Running on ReactNative
+                        const Platform = require('react-native').Platform;
+                        userAgent += Platform.OS + ", v" + Platform.Version;
+                    } else {
+                        // Running on a normal machine
+                        userAgent += process.version;
+                    }
+                    return userAgent += ")";
+                } catch (e) {
+                    return "RealmJS/Unknown"
+                }
+            },
         }));
     }
