+++ lib/index.js
@@ -87,28 +87,6 @@ function getContext() {
     throw Error("Unknown execution context");
 }
 
-function createUserAgentDescription() {
-    // Detect if in ReactNative (running on a phone) or in a Node.js environment
-    // Credit: https://stackoverflow.com/questions/39468022/how-do-i-know-if-my-code-is-running-as-react-native
-    try {
-        var userAgent = "RealmJS/";
-        var context = getContext();
-        userAgent = userAgent + require('../package.json').version + " (" + context + ", ";
-        if (typeof navigator !== 'undefined' && navigator.product === 'ReactNative') {
-            // Running on ReactNative
-            const Platform = require('react-native').Platform;
-            userAgent += Platform.OS + ", v" + Platform.Version;
-        } else {
-            // Running on a normal machine
-            userAgent += process.version;
-        }
-        return userAgent += ")";
-    } catch (e) {
-        return "RealmJS/Unknown"
-    }
-}
-
-
 var realmConstructor;
 
 const context = getContext();
